refactor: ZipFile signed/unsigned cleanup (#120)

+++ lib/Epub/Epub/BookMetadataCache.cpp
@@ -226,8 +226,8 @@ bool BookMetadataCache::buildBookBin(const std::string& epubPath, const BookMeta
     });
 
     spineSizes.resize(spineCount, 0);
-    [[maybe_unused]] const int matched = zip.fillUncompressedSizes(targets, spineSizes);
-    LOG_DBG("BMC", "Batch lookup matched %d/%d spine items", matched, spineCount);
+    [[maybe_unused]] const size_t matched = zip.fillUncompressedSizes(targets, spineSizes);
+    LOG_DBG("BMC", "Batch lookup matched %zu/%d spine items", matched, spineCount);
 
     targets.clear();
     targets.shrink_to_fit();
